refactor: 숙소 팀스페이스 추가 버튼 상세페이지로 이동
